Tried to fix mail problem
Played around with the mailgun - put my own api in etc.

It will send the message if you change 'to' to an actual email address, but it won't if you have 'email' suggesting it is not taking the data from the form!

I have put it back to email for now - but still has my mailgun api in.